axi_dmac: framelock
Move the bypass logic inside the 2d block to ease cascading several
blocks one after other.

scripts/adi_ip.tcl: let interfaces have default value

Allow setting the default value of signals from interfaces during
definition. This is useful when the interface can be left unconnected in
the block design.

axi_dmac: dynamic framelock initial version

This feature adds support for multiple (2d) frame buffers which are used in
a cyclic way. On the same buffers a second DMAC core can operate.

The core can operate in two modes:
  Master mode - available in s2mm configuration, the master (writer) DMAC will
    always skip the current in use slave buffer
  Slave mode - available in mm2s configuration, the slave (reader) DMAC will
    try to stay behind the master buffer by either repeating or skipping
    buffers according to the speed relationship of the two cores.

The master and slave DMAC cores must be connected through the dedicated
'framelock' interface. They must be programmed with similar settings
regarding the buffers size, start address and stride.

The 'framelock' mechanism ensures no buffer is accessed by two DMACs
in the same time.

axi_dmac: support for TLAST marking EndOfLine

In 2D mode, the assertion of TLAST on the AXI Stream source interface
signalizes the end of frame, but Xilinx IP cores use this to signalize
then end of line.

With this change the function of the TLAST will be selectable between
End of Frame and End of Line.

axi_dmac: external sync support

With external synchronization external components to the DMAC  will be able
to throttle the consumption of descriptors queued by the software.

A transfer will start only after the assertion of the external sync
signal. The sync signal can be either in source or destination clock domain or
both.
This feature does not ensures fixed latency from the assertion of
external sync signal and the availability of the data at the destination
interface.

axi_dmac: fix alignment of generate blocks

axi_dmac: support for autorun

In cases where software is not available, the autorun feature can be
used to set up a single transfer that will be executed after reset
deassertion. This is mostly useful together with the CYCLIC mode.

axi_dmac: expose TUSER on AXIS dest interface

TUSER is set for the first beat of the transfer.
In 2D mode will be set only for the first line (transfer)
marking the start of frame.

axi_dmac: simple framelock; wait for master selectable

This change adds support for 'Simple Framelock' mode where the master (writer) DMAC
will not look after the currently read buffer by the slave DMAC, it will
cycle through the buffers regardless of the reader. The slave will always read
a buffer at a predefined distance from the one currently accessed by the master.
This is useful where the writer and reader frame rate match and a predefined
fixed number of frame delay must be produced between reader and writer.

In Simple Framelock mode if the 'wait for master' is enabled the reader will output
a frame only after the master wrote one to the memory. The slave will
follow the master at a defined fixed distance.
If the 'wait for master' is not enabled the slave will start reading a
buffer whenever it completed a previous buffer and receives an external sync
signal if the external synchronization support is enabled.

axi_dmac: rename local parameter

axi_dmac:flock: avoid accessing the same buffer during the first frame

axi_dmac: version bump

This version contains the following features:
  - FrameLock
  - TLAST EOT or EOL
  - External synchronization
  - Autorun

axi_dmac: do not increment measured transfer length during zero completions

Make the measured transfer length more accurate, even if it's a debug
feature.